% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memtools.R
\name{deref}
\alias{deref}
\title{Dereference an object address}
\usage{
deref(x)
}
\arguments{
\item{x}{A string containing a valid memory address of an R object
in hexadecimal format. Can also be a node whose \code{id} field
will be dereferenced.}
}
\value{
The object at the address pointed to by \code{x}.
}
\description{
Dereference an object address
}
