% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{mem_diff}
\alias{mem_diff}
\title{Diff before and after snapshots}
\usage{
mem_diff(before, after)
}
\arguments{
\item{before, after}{Snapshots created by \code{\link[=mem_snapshot]{mem_snapshot()}} before
and after an expression that causes a leak.}
}
\value{
\code{after} without the rows that exist in \code{before}. Note that
the nodes in the \code{node} list-column still refer to nodes in
\code{before} via \code{parents}, \code{children}, etc.
}
\description{
Diff before and after snapshots
}
\examples{
e <- new.env(parent = emptyenv())

e$x <- 1
before <- mem_snapshot(e)

e$y <- 2
after <- mem_snapshot(e)

# The snapshot diff only contains the bucket pairlist node where
# `y` is stored, the symbol `y`, and the double value `2` which is
# bound to `y`
mem_diff(before, after)
}
