% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{mem_igraph}
\alias{mem_igraph}
\title{Add and retrieve igraph structure of snapshot}
\usage{
mem_igraph(snapshot)
}
\arguments{
\item{snapshot}{A memtools snapshot.}
}
\description{
\code{mem_igraph()} retrieves the igraph structure of \code{snapshot}. The
igraph structure is cached inside \code{snapshot} by side effect to
avoid recomputing it multiple times.
}
\details{
The igraph is ordered according to the original data frame
ordering. Rearranging the rows of \code{snapshot} does not rearrange the
igraph.
}
\examples{
s <- mem_snapshot(list(1, list(2)))

# Retrieve the igraph. It is first computed and then cached inside
# the snapshot so it does not need to be recreated again.
g <- mem_igraph(s)

# Work with igraph
igraph::gsize(g)
}
