% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{mem_node_dominator_until}
\alias{mem_node_dominator_until}
\title{Reach dominator of interest}
\usage{
mem_node_dominator_until(.node, .p, ..., .quiet = FALSE)
}
\arguments{
\item{.node}{A memtools node.}

\item{.p}{A predicate function applied to the objects represented
by dominator nodes.}

\item{...}{Arguments passed to \code{.p}.}

\item{.quiet}{Whether to report how many nodes were climbed.}
}
\value{
The dominator node for which \code{.p} returned true. \code{NULL} if
none were found.
}
\description{
\code{mem_node_dominator_ns()} climbs nodes from dominated to dominator
until a namespace is reached.
}
\seealso{
\code{\link[=mem_node_fork]{mem_node_fork()}} and \code{\link[=mem_node_dominator_fork]{mem_node_dominator_fork()}}
}
