% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{mem_node_fork}
\alias{mem_node_fork}
\alias{mem_node_dominator_fork}
\title{Climb nodes until a fork}
\usage{
mem_node_fork(node, quiet = FALSE)

mem_node_dominator_fork(node, quiet = FALSE)
}
\arguments{
\item{node}{A memtools node.}

\item{quiet}{Whether to report how many nodes were climbed.}
}
\description{
\itemize{
\item \code{mem_node_fork()} climbs nodes from child to parent until a node
with more than one parent is reached.
\item \code{mem_node_dominator_fork()} climbs nodes from dominated to
dominator until a node with more than one parent is reached.
}
}
