% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{mem_node_undominated_parents}
\alias{mem_node_undominated_parents}
\alias{mem_node_dominated_ids}
\title{Retrieve dominance properties}
\usage{
mem_node_undominated_parents(node)

mem_node_dominated_ids(node)
}
\arguments{
\item{node}{A memtools node.}
}
\description{
\itemize{
\item \code{mem_node_undominated_parents()} returns the list of parent nodes
which are not in the retained subtree.
\item \code{mem_node_dominated_ids()} returns the ids of all nodes in the
dominated subtree.
}
}
