% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{mem_set_igraph}
\alias{mem_set_igraph}
\alias{mem_igraph}
\title{Add and retrieve igraph structure of snapshot}
\usage{
mem_set_igraph(snapshot)

mem_igraph(snapshot)
}
\arguments{
\item{snapshot}{A memtools snapshot.}
}
\description{
\itemize{
\item \code{mem_set_igraph()} computes the igraph structure of \code{snapshot}. It
is added to the tibble under the \code{mem_igraph} attribute.
\item \code{mem_igraph()} retrieves the igraph structure of \code{snapshot}. If
it has already been computed with \code{mem_set_igraph()} the cached
igraph is returned.
}

The igraph is ordered according to the original data frame
ordering. Rearranging the rows of \code{snapshot} does not rearrange the
igraph.
}
\examples{
s <- mem_snapshot(list(1, list(2)))

# Compute and store the igraph in the snapshot
s <- mem_set_igraph(s)

# Retrieve the igraph
g <- mem_igraph(s)

# Work with igraph
igraph::gsize(g)
}
