% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{mem_shortest_paths}
\alias{mem_shortest_paths}
\title{Find all shortest paths}
\usage{
mem_shortest_paths(snapshot, to, from = NULL)
}
\arguments{
\item{snapshot}{A memory snapshot data frame created by
\code{\link[=mem_snapshot]{mem_snapshot()}}.}

\item{to, from}{Nodes from the \code{node} column of \code{snapshot}.  If
\code{from} is not supplied, the paths from the root are taken by
default. Can also be indices into \code{snapshot$node}.}
}
\description{
Call \code{\link[=mem_set_igraph]{mem_set_igraph()}} on \code{snapshot}
}
