% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{mem_snapshot}
\alias{mem_snapshot}
\title{Create a memory snapshot}
\usage{
mem_snapshot(x)
}
\arguments{
\item{x}{An R object. This becomes the root of the graph of R
objects captured in the snapshot.}
}
\value{
A data frame containing nodes, arrows between these nodes,
and metadata.
}
\description{
Create a memory snapshot
}
\seealso{
\code{\link[=mem_stash]{mem_stash()}} to prevent objects from being recorded in a
snapshot.
}
