% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memtools.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{sexp_address}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rlang}{\code{\link[rlang]{sexp_address}}}
}}

