% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{root_cpp11}
\alias{root_cpp11}
\title{Roots of interest}
\usage{
root_cpp11()
}
\description{
\itemize{
\item \code{root_cpp11()} returns the precious list of cpp11. This is a
doubly linked list preserved in R's own precious list.
}
}
