% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{mem_stash}
\alias{mem_stash}
\title{Stash objects}
\usage{
mem_stash(...)
}
\arguments{
\item{...}{Objects to stash.}
}
\value{
An environment containing the stashed objects. You can add
objects to this environment in the usual fashion.
}
\description{
Objects contained in a stash are never recorded by
\code{\link[=mem_snapshot]{mem_snapshot()}}.
}
